Action Say uses speech synth queue
Action run changed debug.  Action Say now uses the synth generation queue, and handles the callback from it in an invoke to finish it off.
This means the amount of time the system is waiting to generate the speech is not on the main thread, so it runs more smoothly.